Restyle rankings and switch button
