+++ src/components/ranking-chart/ranking-chart-component.jsx
@@ -7,13 +7,9 @@ import PropTypes from 'prop-types';
 import Tooltip from '@tippyjs/react';
 import cx from 'classnames';
 
-// import Dropdown from 'components/dropdown';
-
 import SearchInput from 'components/search-input';
 
 import {
-  // SORT_GROUPS,
-  // getSortOptions,
   getRankingLegend,
   SORT_GROUPS_SLUGS,
 } from 'constants/country-mode-constants';
@@ -24,38 +20,23 @@ import styles from './ranking-chart-styles.module.scss';
 const categories = Object.keys(SORT_GROUPS_SLUGS);
 function RankingChart({
   data,
-  coastal,
   className,
   countryISO,
-  // handleFilterSelection,
-  // selectedFilterOption,
   handleSearchChange,
   handleCountryClick,
   scrollIndex,
   searchTerm,
-  selectedLandMarineOption,
-  handleLandMarineSelection,
 }) {
   const t = useT();
   const locale = useLocale();
   const [hasScrolled, changeHasScrolled] = useState(false);
 
   const rankingLegend = useMemo(() => getRankingLegend(), [locale]);
-  // const sortOptions = useMemo(() => getSortOptions(), [locale]);
   const RANKING_HEADER_LABELS = {
     [SORT_GROUPS_SLUGS.species]: t('species'),
     [SORT_GROUPS_SLUGS.humanModification]: t('human modification'),
     [SORT_GROUPS_SLUGS.protection]: t('protection'),
     [SORT_GROUPS_SLUGS.spi]: t('spi'),
-  };
-
-  const tabsData = {
-    land: {
-      text: t('Land SPI'),
-    },
-    marine: {
-      text: t('Marine SPI'),
-    },
   };
 
   const tableRef = useRef();
@@ -124,38 +105,6 @@ function RankingChart({
     <div className={className}>
       <div className={styles.chartTitleContainer}>
         <span className={styles.chartTitle}>{t('Ranking by')}</span>
-        <div>
-          {Object.keys(tabsData).map((key) => (
-            <button
-              key={key}
-              disabled={!coastal}
-              type="button"
-              className={cx({
-                [styles.switchDataButton]: true,
-                [styles.switchDataActiveButton]:
-                  selectedLandMarineOption.slug === key,
-              })}
-              onClick={() => handleLandMarineSelection(key)}
-            >
-              {tabsData[key].text}
-            </button>
-          ))}
-        </div>
-
-        {/*
-        <span className={cx(styles.chartTitle, styles.filterTitle)}>
-          {t('sort countries')}
-        </span>
-        <div className={styles.dropdownWrapper}>
-          <Dropdown
-            width="full"
-            parentWidth="410px"
-            options={sortOptions}
-            groups={SORT_GROUPS}
-            selectedOption={selectedFilterOption}
-            handleOptionSelection={handleFilterSelection}
-          />
-        </div> */}
       </div>
       <div className={styles.header}>
         <SearchInput
@@ -163,7 +112,6 @@ function RankingChart({
           placeholder={t('Search country')}
           onChange={handleSearchChange}
           value={searchTerm}
-          type="transparent"
         />
         {categories.map((category) => (
           <div
@@ -205,7 +153,6 @@ function RankingChart({
                   <span
                     className={cx(styles.spiCountryIndex, {
                       [styles.found]: scrollIndex === i,
-                      [styles.selectedCountry]: countryISO === d.iso,
                     })}
                   >
                     {`${i + 1}.`}
@@ -213,7 +160,6 @@ function RankingChart({
                   <span
                     className={cx(styles.spiCountryName, {
                       [styles.found]: scrollIndex === i,
-                      [styles.selectedCountry]: countryISO === d.iso,
                     })}
                   >
                     {d.name}
