+++ src/components/search-input-legacy/search-input-component.jsx
