+++ src/components/search-input-legacy/search-input-legacy.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './search-input-component';
