+++ src/pages/nrc/nrc-styles.module.scss
@@ -1,6 +1,7 @@
 @import 'styles/settings';
 @import 'styles/common-animations.module';
 @import 'styles/ui.module';
+@import 'styles/typography-extends';
 
 .background {
   width: 100vw;
@@ -42,13 +43,47 @@
   border-radius: 6px;
 }
 
+.switchDataButtons {
+  position: absolute;
+  top: 18px;
+  right: 20px;
+}
+
+.switchDataButton {
+  @extend %subtitle;
+  border: 1px solid $white;
+  color: $white;
+  padding: 4px 8px;
+  text-transform: capitalize;
+  width: auto;
+  border-top-left-radius: 4px;
+  border-bottom-left-radius: 4px;
+
+  &:last-of-type {
+    border-top-left-radius: 0px;
+    border-bottom-left-radius: 0px;
+    border-top-right-radius: 4px;
+    border-bottom-right-radius: 4px;
+  }
+
+  &:disabled {
+    cursor: not-allowed;
+  }
+}
+
+.switchDataActiveButton {
+  @extend %subtitle;
+  background-color: $white;
+  color: $_temporary-dark-text;
+}
+
 .fullRankingButtonContainer {
   position: absolute;
   top: calc(50% - 140px / 2);
 
   .fullRankingButton {
     width: 24px;
-    height: 150px;
+    height: 160px;
     display: flex;
     flex-direction: column;
     background-color: $black;
@@ -82,7 +117,7 @@
 }
 
 .rankingChart {
-  padding: 16px 24px;
+  padding: 16px 24px 0 0;
   position: relative;
   width: 63vw;
   max-height: 100vh;
