+++ src/styles/override/esri-popup-widget.scss
@@ -67,7 +67,7 @@
 
 .tooltip-area-name {
   letter-spacing: 1.5px;
-  line-height: 14px;
+  line-height: $line-height;
   font-size: 10px;
   text-transform: uppercase;
   margin: 0;
