+++ src/styles/typography-extends.scss
@@ -60,9 +60,9 @@
 %subtitle {
   font-family: $font-family-sans-serif;
   font-size: 12px;
-  font-weight: 300;
+  font-weight: 400;
   line-height: 22px;
-  letter-spacing: 0.02em;
+  letter-spacing: 0.06em;
   text-transform: uppercase;
   @media print {
     font-size: 7pt;
